# swagger.json
{
  "basePath": "/api/v1",
  "definitions": {
...
    "Availability": {
      "properties": {
        "teamId": {
          "format": "uuid",
          "type": "string"
        },
        "type": {
          "enum": [0, 1, 2, 3],
          "type": "integer"
        }
      },
      "type": "object"
    },
    "BackendData": {
      "properties": {
        "name": {
          "example": "my custom backend",
          "type": "string"
        },
        "rest": {
          "example": "https://...",
          "type": "string"
        },
        "ws": {
          "example": "wss://...",
          "type": "string"
        }
      },
      "required": ["name", "rest", "ws"],
      "type": "object"
    },
    "BasicLogin": {
      "properties": {
        "backend": {
          "enum": ["staging", "prod", "production"],
          "type": "string"
        },
        "customBackend": {
          "$ref": "#/definitions/BackendData"
        },
        "email": {
          "example": "email@example.com",
          "format": "email",
          "type": "string"
        },
        "password": {
          "example": "",
          "format": "password",
          "type": "string"
        }
      },
      "required": ["email", "password"],
      "type": "object"
    },
    "Client": {
      "properties": {
        "class": {
          "type": "string"
        },
        "cookie": {
          "type": "string"
        },
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "location": {
          "properties": {
            "lat": {
              "format": "float",
              "type": "integer"
            },
            "lon": {
              "format": "float",
              "type": "integer"
            }
          },
          "required": ["lat", "lon"],
          "type": "object"
        },
        "model": {
          "type": "string"
        },
        "time": {
          "format": "date-time",
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "ErrorMessage": {
      "properties": {
        "code": {
          "type": "integer"
        },
        "error": {
          "type": "string"
        }
      },
      "required": ["code", "error"],
      "type": "object"
    },
...
    "Instance": {
      "allOf": [
        {
          "$ref": "#/definitions/InstanceAndName"
        },
        {
          "properties": {
            "backend": {
              "type": "string"
            },
            "clientId": {
              "type": "string"
            }
          },
          "type": "object"
        }
      ]
    },
    "InstanceAndName": {
      "properties": {
        "instanceId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "Login": {
      "allOf": [
        {
          "$ref": "#/definitions/BasicLogin"
        },
        {
          "properties": {
            "deviceClass": {
              "enum": ["desktop", "phone", "tablet"],
              "type": "string"
            },
            "deviceLabel": {
              "example": "",
              "type": "string"
            },
            "deviceName": {
              "example": "",
              "type": "string"
            },
            "name": {
              "type": "string"
            }
          },
          "type": "object"
        }
      ],
      "description": "**Note**: You can either set `backend` or `customBackend`. If you set neither, the \"staging\" backend will be used. If you set both, `backend` takes the precedence.",
      "required": ["email", "password"]
    },
...
    "ServerErrorMessage": {
      "allOf": [
        {
          "$ref": "#/definitions/ErrorMessage"
        },
        {
          "properties": {
            "stack": {
              "type": "string"
            }
          },
          "type": "object"
        }
      ]
    },
...
  },
  "info": {
    "contact": {
      "email": "opensource@wire.com"
    },
    "description": "End-to-end Test Service (ETS) for Wire's test automation suite.",
    "license": {
      "name": "GPL-3.0",
      "url": "https://github.com/wireapp/wire-web-ets/blob/master/LICENSE"
    },
    "termsOfService": "https://wire.com/legal/",
    "title": "E2E Test Service",
    "version": "1.20.0"
  },
  "paths": {
...
    "/clients": {
      "delete": {
        "consumes": ["application/json"],
        "description": "**Note**: You can either set `backend` or `customBackend`. If you set neither, the \"staging\" backend will be used. If you set both, `backend` takes the precedence.",
        "operationId": "deleteAllClients",
        "parameters": [
          {
            "description": "Login data",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BasicLogin"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "items": {
                "$ref": "#/definitions/Client"
              },
              "type": "array"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Delete all clients",
        "tags": ["Clients"]
      }
    },
...
    "/instance": {
      "put": {
        "consumes": ["application/json"],
        "description": "",
        "operationId": "createInstance",
        "parameters": [
          {
            "description": "Login data",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/InstanceAndName"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Create a new instance",
        "tags": ["Instance"]
      }
    },
    "/instance/{instanceId}": {
      "delete": {
        "description": "",
        "operationId": "deleteInstance",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Delete an instance",
        "tags": ["Instance"]
      },
      "get": {
        "operationId": "getInstance",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Instance"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Get information about an instance",
        "tags": ["Instance"]
      }
    },
...
    "/instance/{instanceId}/availability": {
      "post": {
        "operationId": "setAvailability",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          },
          {
            "description": "Type can be `0` (none), `1` (available), `2` (away), `3` (busy).",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Availability"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/InstanceAndName"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Set a user's availability",
        "tags": ["Instance"]
      }
    },
...
    "/instance/{instanceId}/clients": {
      "get": {
        "operationId": "getClients",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "items": {
                "$ref": "#/definitions/Client"
              },
              "type": "array"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Get all clients of an instance",
        "tags": ["Instance"]
      }
    },
...
    "/instances": {
      "get": {
        "operationId": "getInstances",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Instance"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Get all instances",
        "tags": ["Instances"]
      }
    },
...
  },
  "schemes": ["http"],
  "swagger": "2.0",
...
}